Changed, podspec name to DBAccess from DB-Access
Upgraded to v1.6.4:

FIX:  Bug where the joinedResults dictionary was not getting populated.
FIX:  Bug with initWithPrimaryKey:, this was working for existing objects but failing to create a new class for objects which were not a match.
Support property names that start with an uppercase letter.
Added support for Default property values.
Allow reserved SQLITE words to be used as property (column) names, such as where, as, default.

Signed-off-by: Adrian Herridge <adrian@ipresent.com>